refactor: put context in one folder
