+++ src/components/pages/ExplorerPage/ExplorerPage.tsx
@@ -18,7 +18,7 @@ import { useState, useEffect } from "react";
 
 import ListView from "./ListView";
 import DetailsView from "./DetailsView";
-import { useAuth } from "context/AuthContext";
+import { useAuth } from "context";
 import { Project } from "types";
 
 const ExplorerPage: React.FC = () => {
